pyproject.toml
[tool.poetry]
name = "frontend"
version = "0.1.0"
description = ""
authors = ["mdfarhanali34 <farhandxb96@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.41.1"
python-dotenv = "^1.0.1"
httpx = "^0.28.1"
pydantic = "^2.10.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
